backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<0.29.0
sqlalchemy>=2.0.0,<3.0.0
pydantic-settings>=2.2.1,<3.0.0
python-multipart>=0.0.9,<0.0.10
pypdf>=4.0.0,<5.0.0
langchain>=0.1.13,<0.2.0
chromadb>=0.4.24,<0.5.0
openai>=1.14.0,<2.0.0
google-generativeai>=0.5.0,<1.0.0
tenacity>=8.2.0,<9.0.0
httpx>=0.26.0,<0.27.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<0.24.0
